Initial version of PWM class
To activate PWM pin, we're using these 2 command:
echo am33xx_pwm > /sys/devices/bone_capemgr.?/slots
cd /sys/devices/ocp.?/pwm_test_P8_13.??/

Since using wildcard in open() function is not possible, we have
to know which numbers that we'll use in place of '?'.

This version of class, can find the relevant number that we need.
And it can load am33xx_pwm firmware. That's all for now.